refactor appointment date provider and filter strategies
