+++ packages/devextreme/js/__internal/scheduler/r1/utils/index.ts
@@ -47,7 +47,6 @@ export {
   extendGroupItemsForGroupingByDate,
   getAppointmentKey,
   getAppointmentRenderingStrategyName,
-  getAppointmentTakesAllDay,
   getCalculatedFirstDayOfWeek,
   getCellDuration,
   getDatesWithoutTime,
@@ -71,6 +70,7 @@ export {
   getViewStartByOptions,
   getWeekendsCount,
   hasResourceValue,
+  isAppointmentTakesAllDay,
   isDataOnWeekend,
   isDateAndTimeView,
   isDateInRange,
